+++ tests/testthat/test_AIC.R
@@ -6,6 +6,7 @@ test_that("The AIC is reproducible", {
                     data.frame(df = c(3, 4, 5, 5), AIC = c(59.3, 44.7, 29.0, 39.2)),
                     scale = 1, tolerance = 0.1)
   expect_error(AIC(fits["SFO", ]), "column object")
+  expect_error(BIC(fits["SFO", ]), "column object")
   expect_equivalent(BIC(fits[, "FOCUS_C"]),
                     data.frame(df = c(3, 4, 5, 5), AIC = c(59.9, 45.5, 30.0, 40.2)),
                     scale = 1, tolerance = 0.1)