+++ tests/testthat/test_plot.R
@@ -2,7 +2,6 @@ context("Plotting")
 
 test_that("Plotting mkinfit and mmkin objects is reproducible", {
   skip_on_cran()
-  skip_on_travis()
   plot_default_FOCUS_C_SFO <- function() plot(fits[["SFO", "FOCUS_C"]])
   plot_res_FOCUS_C_SFO <- function() plot(fits[["SFO", "FOCUS_C"]], show_residuals = TRUE)
   plot_res_FOCUS_C_SFO_2 <- function() plot_res(fits[["SFO", "FOCUS_C"]])
@@ -16,6 +15,11 @@ test_that("Plotting mkinfit and mmkin objects is reproducible", {
 
   plot_errmod_fit_D_obs_eigen <- function() plot_err(fit_D_obs_eigen, sep_obs = FALSE)
   plot_errmod_fit_C_tc <- function() plot_err(fit_C_tc)
+
+  plot_biphasic_mmkin <- function() plot(mixed(mmkin_biphasic))
+  plot_biphasic_nlme <- function() plot(nlme_biphasic)
+  plot_biphasic_saem_s <- function() plot(saem_biphasic_s)
+  plot_biphasic_saem_m <- function() plot(saem_biphasic_m)
 
   plot_res_sfo_sfo <- function() plot_res(f_sfo_sfo_desolve)
   plot_err_sfo_sfo <- function() plot_err(f_sfo_sfo_desolve)
@@ -32,6 +36,10 @@ test_that("Plotting mkinfit and mmkin objects is reproducible", {
   vdiffr::expect_doppelganger("mmkin plot for FOCUS C", mmkin_FOCUS_C)
   vdiffr::expect_doppelganger("mmkin plot for SFO (FOCUS C and D)", mmkin_SFO)
   vdiffr::expect_doppelganger("plot_errmod with FOCUS C tc", plot_errmod_fit_C_tc)
+  vdiffr::expect_doppelganger("mixed model fit for mmkin object", plot_biphasic_mmkin)
+  vdiffr::expect_doppelganger("mixed model fit for nlme object", plot_biphasic_nlme)
+  vdiffr::expect_doppelganger("mixed model fit for saem object with saemix transformations", plot_biphasic_saem_s)
+  vdiffr::expect_doppelganger("mixed model fit for saem object with mkin transformations", plot_biphasic_saem_m)
   skip_on_travis() # Still not working on Travis, maybe because of deSolve producing
   # different results when not working with a compiled model or eigenvalues
   vdiffr::expect_doppelganger("plot_errmod with FOCUS D obs eigen", plot_errmod_fit_D_obs_eigen)